chore: recuparar perfil do paciente

+++ src/main/java/com/hdc/hdc/pacientes/PacienteRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.Optional;
 
-public interface PacienteRepository extends JpaRepository<Paciente, Long> {
+public interface PacienteRepository extends JpaRepository<Paciente, Integer> {
     boolean existsByEmail(String email);
 
     boolean existsByUsername(String username);
